+++ HISTORY.rst
@@ -10,6 +10,8 @@ Unreleased
 
 **Bugfixes**
 
+* Card 137: Fix group allow to change sale order line price unit
+
 **Build**
 
 **Documentation**